+++ src/api/auth/auth.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { GoogleStrategy } from 'src/common/strategies/google.strategy';
 import { PrismaService } from 'src/prisma.service';
-import { SessionSerializer } from './Serializer';
+import { SessionSerializer } from '../../common/serializer/session.serializer';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
